refactor: deduplicate git cloning and token logic in repomanager
