+++ internal/server/handler/webui.go
@@ -7,7 +7,6 @@ import (
 	"fmt"
 	"log/slog"
 	"net/http"
-	"os"
 	"path/filepath"
 	"strings"
 	"time"
@@ -16,7 +15,6 @@ import (
 
 	"github.com/sevigo/code-warden/internal/config"
 	"github.com/sevigo/code-warden/internal/core"
-	"github.com/sevigo/code-warden/internal/github"
 	"github.com/sevigo/code-warden/internal/gitutil"
 	"github.com/sevigo/code-warden/internal/rag"
 	indexpkg "github.com/sevigo/code-warden/internal/rag/index"
@@ -250,10 +248,6 @@ func (h *WebUIHandler) GetScanStatus(w http.ResponseWriter, r *http.Request) {
 	h.json(w, toScanStateResponse(state))
 }
 
-func isPlaceholderToken(token string) bool {
-	return token == "" || strings.HasPrefix(token, "ghp_your_")
-}
-
 func (h *WebUIHandler) TriggerScan(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	repoID, err := parseRepoID(r)
@@ -304,54 +298,16 @@ func (h *WebUIHandler) runScan(repoID int64, repo *storage.Repository) {
 	}
 }
 
-func (h *WebUIHandler) ensureRepoCloned(ctx context.Context, repo *storage.Repository) error {
-	cleanPath := filepath.Clean(repo.ClonePath)
-	_, err := os.Stat(cleanPath)
-	if err == nil {
-		return nil
-	}
-	if !os.IsNotExist(err) {
-		return fmt.Errorf("stat clone path: %w", err)
-	}
-
-	h.logger.Info("Repository not found locally, cloning", "repo", repo.FullName, "path", cleanPath)
-
-	// 1. Determine which token to use
-	token := h.cfg.GitHub.Token
-	if isPlaceholderToken(token) {
-		token = ""
-	}
-	if repo.InstallationID > 0 {
-		_, instToken, err := github.CreateInstallationClient(ctx, h.cfg, repo.InstallationID, h.logger)
-		if err != nil {
-			h.logger.Warn("failed to create installation client for manual scan, falling back to config token",
-				"repo", repo.FullName,
-				"installation_id", repo.InstallationID,
-				"error", err)
-		} else {
-			token = instToken
-		}
-	}
-
-	cloneURL := fmt.Sprintf("https://github.com/%s.git", repo.FullName)
-	if err := os.MkdirAll(filepath.Dir(cleanPath), 0750); err != nil {
-		return fmt.Errorf("create parent dir: %w", err)
-	}
-	if _, err := h.gitClient.Clone(ctx, cloneURL, cleanPath, token); err != nil {
-		return fmt.Errorf("clone repository: %w", err)
-	}
-	return nil
-}
-
 func (h *WebUIHandler) doScan(ctx context.Context, repoID int64, repo *storage.Repository) error {
-	// Check if repo exists locally, clone if needed
-	if err := h.ensureRepoCloned(ctx, repo); err != nil {
-		return err
+	ev := &core.GitHubEvent{
+		RepoFullName:   repo.FullName,
+		RepoCloneURL:   fmt.Sprintf("https://github.com/%s.git", repo.FullName),
+		InstallationID: repo.InstallationID,
 	}
 
-	scanResult, err := h.repoMgr.ScanLocalRepo(ctx, repo.ClonePath, repo.FullName, true)
+	scanResult, err := h.repoMgr.SyncRepo(ctx, ev, "")
 	if err != nil {
-		return fmt.Errorf("scan local repo: %w", err)
+		return fmt.Errorf("sync repo: %w", err)
 	}
 
 	repoConfig, err := config.LoadRepoConfig(scanResult.RepoPath)
